+++ Cargo.toml
@@ -33,5 +33,8 @@ reqwest = "0.11.10"
 qrcode-png = "0.4.0"
 typed-builder = "0.10"
 
+bincode = "1.3.3" 
+base64 = "0.13.0"
+
 [profile.release]
 lto = true
